Move arrow tests closer to where they are implemented
