+++ boa/src/syntax/parser/expression/assignment/arrow_function/tests.rs
@@ -1,0 +1,295 @@
+use crate::syntax::{
+    ast::node::{
+        ArrowFunctionDecl, BinOp, FormalParameter, Identifier, LetDecl, LetDeclList, Node, Return,
+    },
+    ast::op::NumOp,
+    parser::tests::check_parser,
+};
+
+/// Checks an arrow function with only a rest parameter.
+#[test]
+fn check_arrow_only_rest() {
+    check_parser(
+        "(...a) => {}",
+        vec![ArrowFunctionDecl::new(vec![FormalParameter::new("a", None, true)], vec![]).into()],
+    );
+}
+
+/// Checks an arrow function with a rest parameter.
+#[test]
+fn check_arrow_rest() {
+    check_parser(
+        "(a, b, ...c) => {}",
+        vec![ArrowFunctionDecl::new(
+            vec![
+                FormalParameter::new("a", None, false),
+                FormalParameter::new("b", None, false),
+                FormalParameter::new("c", None, true),
+            ],
+            vec![],
+        )
+        .into()],
+    );
+}
+
+/// Checks an arrow function with expression return.
+#[test]
+fn check_arrow() {
+    check_parser(
+        "(a, b) => { return a + b; }",
+        vec![ArrowFunctionDecl::new(
+            vec![
+                FormalParameter::new("a", None, false),
+                FormalParameter::new("b", None, false),
+            ],
+            vec![Return::new(
+                BinOp::new(NumOp::Add, Identifier::from("a"), Identifier::from("b")),
+                None,
+            )
+            .into()],
+        )
+        .into()],
+    );
+}
+
+/// Checks an arrow function with expression return and automatic semicolon insertion
+#[test]
+fn check_arrow_semicolon_insertion() {
+    check_parser(
+        "(a, b) => { return a + b }",
+        vec![ArrowFunctionDecl::new(
+            vec![
+                FormalParameter::new("a", None, false),
+                FormalParameter::new("b", None, false),
+            ],
+            vec![Return::new(
+                BinOp::new(NumOp::Add, Identifier::from("a"), Identifier::from("b")),
+                None,
+            )
+            .into()],
+        )
+        .into()],
+    );
+}
+
+/// Checks arrow function with empty return
+#[test]
+fn check_arrow_epty_return() {
+    check_parser(
+        "(a, b) => { return; }",
+        vec![ArrowFunctionDecl::new(
+            vec![
+                FormalParameter::new("a", None, false),
+                FormalParameter::new("b", None, false),
+            ],
+            vec![Return::new::<Node, Option<_>, Option<_>>(None, None).into()],
+        )
+        .into()],
+    );
+}
+
+/// Checks an arrow function with empty return, with automatic semicolon insertion.
+#[test]
+fn check_arrow_empty_return_semicolon_insertion() {
+    check_parser(
+        "(a, b) => { return }",
+        vec![ArrowFunctionDecl::new(
+            vec![
+                FormalParameter::new("a", None, false),
+                FormalParameter::new("b", None, false),
+            ],
+            vec![Return::new::<Node, Option<_>, Option<_>>(None, None).into()],
+        )
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment() {
+    check_parser(
+        "let foo = (a) => { return a };",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![FormalParameter::new("a", None, false)],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_nobrackets() {
+    check_parser(
+        "let foo = (a) => a;",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![FormalParameter::new("a", None, false)],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_noparenthesis() {
+    check_parser(
+        "let foo = a => { return a };",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![FormalParameter::new("a", None, false)],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_noparenthesis_nobrackets() {
+    check_parser(
+        "let foo = a => a;",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![FormalParameter::new("a", None, false)],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_2arg() {
+    check_parser(
+        "let foo = (a, b) => { return a };",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![
+                        FormalParameter::new("a", None, false),
+                        FormalParameter::new("b", None, false),
+                    ],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_2arg_nobrackets() {
+    check_parser(
+        "let foo = (a, b) => a;",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![
+                        FormalParameter::new("a", None, false),
+                        FormalParameter::new("b", None, false),
+                    ],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_3arg() {
+    check_parser(
+        "let foo = (a, b, c) => { return a };",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![
+                        FormalParameter::new("a", None, false),
+                        FormalParameter::new("b", None, false),
+                        FormalParameter::new("c", None, false),
+                    ],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
+
+#[test]
+fn check_arrow_assignment_3arg_nobrackets() {
+    check_parser(
+        "let foo = (a, b, c) => a;",
+        vec![LetDeclList::from(vec![LetDecl::new::<_, Option<Node>>(
+            Identifier::from("foo"),
+            Some(
+                ArrowFunctionDecl::new(
+                    vec![
+                        FormalParameter::new("a", None, false),
+                        FormalParameter::new("b", None, false),
+                        FormalParameter::new("c", None, false),
+                    ],
+                    vec![Return::new::<Node, Option<_>, Option<_>>(
+                        Some(Identifier::from("a").into()),
+                        None,
+                    )
+                    .into()],
+                )
+                .into(),
+            ),
+        )])
+        .into()],
+    );
+}
